Move `ChainWorkerWithAttemptedChanges` type
Place it in its own module.
